requirements.txt
click==7.1.1
Flask==1.1.1
Flask-API==2.0
flask-mongoengine==0.9.5
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
mongoengine==0.19.1
PyJWT==1.7.1
pymongo==3.10.1
six==1.14.0
Werkzeug==1.0.1
WTForms==2.2.1
